requirements.txt
# BzR CTM - Backend Django
django>=5.2
djangorestframework>=3.15
django-cors-headers>=4.4
anthropic>=0.40
pdfplumber>=0.11

# Para produção com PostGIS (descomente):
# psycopg2-binary>=2.9

# Flask legado (BCI reader original)
flask>=3.1
werkzeug>=3.1
